Add projects to the sprint page
Added grid with projects logos, links, descrptions, etc
to the sprint page.
Some project descriptions were adapted from the original submission
to have a similar amount of text for each description.

The section related the submission was removed, and the one
to participate was modified to point to the eventbrite site
for people to get tickets.

The 'Propose a sprint' link on the main menu was moved to the "Program"
section, under the 'Sprint projects' name.